requirements.txt
firebase
pyserial